feat(install): Install scripts replace API token in supplied config
Modify regex to include API tokens in user supplied config, when
  replacing config api_token value with the argument provided,
  in the installer scripts.

issue #CIRC-8723

This will replace the existing api_token in the config with the one
passed to the installer, and gives a better user experience when using
the installer. If a customer were to pass around an existing config for
usage on different systems, this gives more predictable behavior, which
is defaulting to the api token that is passed as an argument at runtime.

* Tags: installer macos linux windows freebsd