fix(internal/circonus): All plugins can update check bundle tags
Fix NewMetricDestination to rectify tags for all plugins, not just the
host plugin

issue #CIRC-8707

This changes the way that check bundles are created/maintained to
automatically rectify tags for all plugins, not just the host plugin.
Should we decide to make tags configurable per plugin, this will also
facilitate that.

* Tags: circonus agent input plugin check bundle